pyproject.toml
[project]
name = "forgeagentcoder"
version = "0.1.0"
description = "AWS-native fullstack agentic coding model adaptation pipeline."
requires-python = ">=3.10,<3.12"
dependencies = [
  "boto3>=1.34",
  "botocore>=1.34",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "typer>=0.12",
  "rich>=13.7",
  "tqdm>=4.66",
  "numpy>=1.26",
  "pandas>=2.2",
  "datasets>=2.20",
  "huggingface_hub>=0.24",
  "transformers>=4.44",
  "accelerate>=0.33"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
  "mypy>=1.10"
]

gpu = [
  "peft>=0.12",
  "trl>=0.9",
  "bitsandbytes>=0.43",
  "sentencepiece>=0.2",
  "safetensors>=0.4"
]

[project.scripts]
forgeagent = "forgeagentcoder.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
